backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
PyJWT
python-dotenv
